# Replace symmetric_difference() with difference()