--- a/audformat/utils/__init__.py
+++ b/audformat/utils/__init__.py
@@ -1,5 +1,6 @@
 from audformat.core.utils import (
     concat,
+    difference,
     duration,
     expand_file_path,
     hash,
@@ -15,7 +16,6 @@ from audformat.core.utils import (
     read_csv,
     replace_file_extension,
     set_index_dtypes,
-    symmetric_difference,
     to_filewise_index,
     to_segmented_index,
     union,